fix: 相机模块
相机模块

影响模块：

+++ JODevelop/Test/Transition/PopUpTwoVC.m
@@ -74,7 +74,7 @@
         cell.selectionStyle = UITableViewCellSelectionStyleNone;
     }
     
-    cell.textLabel.text = [NSString stringWithFormat:@"%d",indexPath.row];
+    cell.textLabel.text = [NSString stringWithFormat:@"%ld",(long)indexPath.row];
     return cell;
 }
